+++ monolib-dart/lib/src/pillar/pillar_change_notifier_internal.dart
@@ -1,0 +1,7 @@
+import 'pillar_change_notifier.dart';
+
+class PillarChangeNotifierInternal extends PillarChangeNotifier {
+  void notify() {
+    notifyListeners();
+  }
+}