+++ monolib-flutter/lib/src/pillar/on_pillar_listenable.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/foundation.dart';
+import 'package:monolib_dart/pillar.dart';
+
+import 'pillar_listenable_adapter.dart';
+
+extension OnPillarListenable on PillarListenable {
+  Listenable get asListenable => PillarListenableAdapter(this);
+}